Page: Make use of new Head component
