+++ components/page/package.json
@@ -28,6 +28,7 @@
     "@not-govuk/breadcrumbs": "workspace:^0.14.0",
     "@not-govuk/component-helpers": "workspace:^0.14.0",
     "@not-govuk/footer": "workspace:^0.14.0",
+    "@not-govuk/head": "workspace:^0.14.0",
     "@not-govuk/header": "workspace:^0.14.0",
     "@not-govuk/link": "workspace:^0.14.0",
     "@not-govuk/phase-banner": "workspace:^0.14.0",
