+++ components/page/src/NotGovUKPage.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { FC, createElement as h } from 'react';
-import reactHelmetDefault, * as reactHelmetNamed from 'react-helmet-async';
+import { Head } from '@not-govuk/head';
 import { Page, PageProps } from './Page';
 
 import favicon from '../assets/coat-favicon.ico';
@@ -13,8 +13,6 @@ import ogImage from '../assets/coat-opengraph-image.png';
 
 import '../assets/NotGovUKPage.scss';
 
-const reactHelmet = reactHelmetDefault || reactHelmetNamed;
-const { Helmet } = reactHelmet;
 
 export type NotGovUKPageProps = Omit<PageProps, 'govUK'>;
 
@@ -24,7 +22,7 @@ export const NotGovUKPage: FC<NotGovUKPageProps> = ({ children, classModifiers, 
     classModifiers={[ ...(Array.isArray(classModifiers) ? classModifiers : [classModifiers]), 'not-govuk' ]}
     govUK={false}
   >
-    <Helmet>
+    <Head>
       <meta name="theme-color" content="#0b0c0c" />
       <link rel="shortcut icon" sizes="16x16 32x32 48x48" href={favicon} type="image/x-icon" />
       <link rel="apple-touch-icon" sizes="180x180" href={appleTouchIcon180} />
@@ -32,7 +30,7 @@ export const NotGovUKPage: FC<NotGovUKPageProps> = ({ children, classModifiers, 
       <link rel="apple-touch-icon" sizes="152x152" href={appleTouchIcon152} />
       <link rel="apple-touch-icon" href={appleTouchIcon} />
       <meta property="og:image" content={ogImage} />
-    </Helmet>
+    </Head>
     {children}
   </Page>
 );
