+++ components/page/src/index.ts
@@ -1,7 +1,6 @@
 import Page from './Page';
 
 export default Page;
-export { SimplePage } from './SimplePage';
 export { GovUKPage } from './GovUKPage';
 export { NotGovUKPage } from './NotGovUKPage';
 export * from './Page';
